move tests test_blocks and ruff
